Remove dependency on gitective
gitective is built against a version of jgit which is older than the
one in Gerrit, and uses the release() method which does not exist any
more.  This causes NoSuchMethodError exceptions at runtime.

gitective is only being used to open the OWNERS file in the repository,
so remove it and replace the implementation with direct usage of the
jgit API.

Change-Id: I10c1679da0d39baf04d013fa9579f85140e2b8c4