Update patch set 1
Patch Set 1:

Thanks luca and david for the various feedbacks.
I think that the probable best path for this change to go on is to
verify that under the gerrit v2.13 stable plugin compiles correctly with the change to findall (which now is not compiling because of the final thing). At that point I can abandon this change and insert a new one with just the gitective upgrade otherwise a nasty runtime error happens when plugin scans for the paths. What do you think?

Patch-set: 1