Update patch set 2
Patch Set 2:

Hi! I just updated the commit message, and pushed to master. We would need this patch on 2.11 - as of now the plugin crashes at runtime with 2.11, which embeds a version of jgit which does not come with an API used by the version of gitective used by the plugin in the sable-2.11 branch..

Patch-set: 2