Move to Fabric8 Gitective with JGit 4.x support
Gerrit 2.11.x is based on a JGit 4.x API which is incompatible
with the previous version of Gitective.

The Fabric8 fork has the support for the new API and is actively
maintained.

A quick-fix for this branch is thus just move to Fabric8 to resume
the full functionality of the plugin.

Change-Id: I72c556aff217bb04f4b30bd83fe57049db51788a