[WIP] Refactoring for matching based rules in OWNERS config
Refactoring of some classes to be able to set up some unit tests, and first layout
for hosting path/matching information to extract the owners based on the name files.
An example of the new OWNERS syntax:

inherits: false
owners:
- email1
- email2
matches:
- regex: .*.scala
  owners:
  - email3
  - email4
- exact: Product.java
  owners:
  - email5
  - email6
- suffix: .java
  owners:
  - email7
  - email8

This will be parsed into an additional Map<String,Matcher>matchers where the key is
the full partial path where the OWNERS is found with appended the regex/exact/suffix name part,
for instance if OWNERS is in /project/OWNERS then we will have the following content:
/project/.*.scala containing RegexMatcher instance
/project/Product.java containing ExactMatcher instance
/project/.java containing SuffixMatcher instance

XXXXMatcher instance contains the information about the original string (without partial path)
and a set of Account.Id associated with the owner.

Maps will allow Prolog or other java functions to do quick lookup for single files and finding
the full list of owners.

NOTE: developed to be backward compatible with old OWNERS syntax and only add predicates
and functionalities.
Still not written the real prolog of java part which makes it active, and still under testing.
Further testing or discussions might need further refactoring of this structure.

NOTE: java8 stream procedures have been used in some places

Change-Id: If5a389ad696bdec236f68af4bf1ee65f1d6d04fa