FileOwners added and UnitTested
Allow the plugin storing the full list of owners
responsible for each modified path. Previous version didn't
store this information assuming the only information needed
was that obtained from each matching rule. 
It is far more useful to have this information stored explicitly
for each file.

With this information we can with a future change return this
information to prolog rules, allowing them to understand
for each file if owners are missing.

Prepare the map and unit test it so to be sure it is properly
loaded with standard and extended configuration.

Change-Id: I4de7bac2030474e51f982c61611b7d0dc14ab5a8