FileOwners added and UnitTested
This change allows the plugin to store the full list of owners
responsible for each modified path. Previous version didn't
store this information assuming the only information needed
was that obtained from each matching rule, but eventually
we acknowledged that it is far more useful to have this
information stored explicitly for each file.

With this information we can with a future change return this
information to prolog rules, allowing them to understand
for each file if owners are missing.

This change is so only specifically prepare the map and
unit test it so to be sure it is properly loaded with
standard configuration and with extended configuration.

Change-Id: I4de7bac2030474e51f982c61611b7d0dc14ab5a8