Bazel: De-duplicate plugins API in final artifact
The build implementation currently erroneously transitively depends on
the plugin API. As a consequence, the size of final artifact is 50 MB.

After fixing it, it turns out that the transitive dependency on
prolog:common rule in Gerrit core pulls in another 45 MB dependency from
Gerrit core. That is due to the fact, that prolog:common transitively
depends on gerrit-server that transitively depends on the virtually
the whole Gerrit build graph.

Fix it by exposing the missing prolog rule in the plugin API, and cut
off the prolog runtime dependency as well. Prolog runtime is already
included in the Gerrit final artifact anyway.

Change-Id: If9d249fb9bb2e67a6103159157fd7fc5d593d7cb