Adapt retrieval of patch-list to Gerrit master
Because of the refactoring of PatchListKey on Gerrit master
the owner-autoassign need to change the way it fetches the list
of patches associated to a new commit, using the new
static methods provided.

The old direct creation of PatchListKey is not an option anymore
because the constructor has been restricted to package level access.

Change-Id: I4349ed6b20ccaee6f9934928c8aa4b18f87f2b78