Sync Jackson lib version with Gerrit
When trying to implement standalone build, it was found that plugin was
including the jackson-core artifact defined in Gerrit and not the one
defined in the plugin. Update jackson dependencies to match the version
defined in Gerrit.

Given jackson-core is already loaded by Gerrit, avoid loading it when
building in-tree. The dependency is kept, however, in order to allow
standalone build (in follow-up change) to work.

Change-Id: I6c06f9ce0a7338c614e5833ac0de9084db9acbb7