Move root package to com.googlesource
All the plugins hosted on gerrit-review.googlesource.com should be
based on the com.googlesource. root package, to indicate that they
are not owned or maintained by a single company but the Gerrit
Community.

Having exceptions gives the green light to more plugins with vendors'
or companies' names which is not desirable.

Git does a very good job in merging renamed files, this mass rename
should not create then any problems.

Change-Id: If0cb595050484bb72e70bdfe2456adb85b7f02d3