Update patch set 1
Patch Set 1:

> Patch Set 1: Verified-1
> 
> > Patch Set 1:
> > 
> > > Patch Set 1: Verified-1
> > > 
> > > 'bazel build :all' fails
> > 
> > Never used the 'all' target on this plugin. Not sure how it would work in tree-mode.
> > This plugin is actually two different plugins:
> > - owners
> > - owners-autoassign
> > 
> > The CI build doesn't use that target at all. Do you guys use it?
> 
> We at least used to have it (:all) used in our CI job so far -not sure why. This change fails on both :all and 'owners' targets for me locally. This change base 6c0cc33 (current stable-2.15) also has :all failing for me, btw; however 'bazel build owners' works on that base. And current stable-2.14 has both targets passing. -Just my humble feedback/vote then.

I tried to build as we typically do on gerrit-ci.googlesource.com and it is all fine (we use the in-tree mode).

Can you create an issue on the v2.15 on the standalone Bazel build for the owners and let this go through? It isn't introduced by this change after all.

Patch-set: 1
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>