Bazel: Fix standalone build
In recent bazlets version, stamp generation was extended to grep not
only for the plugin name, but for STABLE_BUILD_<plugin-name>_LABEL.
For this reason, the grep for both OWNERS and OWNERS-AUTOASSIGN does
not work any more.

Add another line to the generated bazel-out/stable-status.txt file to
rectify. This problem is very special, because two plugins with
different names co-exist in the same repository.

Test Plan:

  $ bazel build :all

Bug: Issue 9945
Change-Id: Idfe231cbe12d0c58ef5208044140a9f56adf9a49