Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Unless this is a v2.15/v2.16 regression, I would rather make the documentation to reflect the backward compatible behaviour.

I checked the source code, and I think this bug was introduced by 
https://gerrit.googlesource.com/plugins/owners/+/7c322d75aaa3447450c546ab22ac56da5e4c9263%5E%21/#F0,  which I guess it's in v2.12 or v2.13.

I agree it might a bit risky to fix this, but I think it is more natural and reasonable to inherit ownerships by default practically.

If you think this is too risky, what about fixing this in v2.16? 
I know you're working on a new owners plugin without prolog, and so I think it might be a good time to fix this bug.

Patch-set: 1
Reviewer: Gerrit User 1047142 <1047142@173816e5-2b9a-37c3-8a2e-48639d4f1153>