Prevent infinite event processing when notedb is enabled
When note db is enabled, the assignment of owners to a particular change
might update the meta refs, which in turn would trigger the processing of
a new GitReferenceUpdate event, and so on.

This change prevents this situation by making sure that meta refs
updates are not processed by the GitRefListener.

Bug: Issue 9756
Change-Id: Ief0a82b877d7b1446ba2b9bf04f9f25a2cc866b2