Create a user-context if not propagated by Gerrit
Because of (Bug: Issue 10941), the user context could be
absent and thus all the subsequent operations performed on
the change would fail.

Provide a one-off context using the updater's identity
so that even when Gerrit isn't providing a valid context
the overall OWNERS processing logic would still succeed.
Include a new test to make sure that the scenario is
covered from now on.

Change-Id: If9f2087519180baddd996f41bdda56942b3f6886