Update patch set 4
Patch Set 4:

Thanks, DavidO, that is helpful.
The boilerplate for standalone builds, as Bazel moves on, is getting more and more complicated :-(

Have you spoken to them about the problem in Munich? Are we doing something very uncommon here? Are the standalone builds a supported use-case for the Bazel Team?

Patch-set: 4