Do not assign reviewers that cannot see the change
When assigning reviewers to a change automatically, do not
include the ones that have no read visibility to the change
ref. Trying to assign users that cannot see the change would
result in throwing an exception and would thus break the entire
functionality of auto-assigning reviewers to a change.

Keep the message of the users not assigned to a change
as a warning so that the Gerrit admin could understand the problem
and adjust the ACLs if needed.

Bug: Issue 12118
Change-Id: Id84d10d13cb3a5291d7275b2ad24e4686464380d