Update patch set 1
Patch Set 1:

Fixed with https://gerrit-review.googlesource.com/c/plugins/owners/+/249386.
I do prefer to keep the current merge as-is and fix the functionality in the follow-up change.

It would allow to have a clear view on the merge and the adaptation to v2.15.

Patch-set: 1
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>