Limit the initial attention-set to 2 reviewers
Avoid adding too many reviewers automatically to the attention-set
when a change is created. The OWNERS list, because of hierarchy, may
pull a lot of names as reviewers, but that doesn't mean that all of them
should have the initial task to provide the first feedback.

Only two random 2 reviewers are assigned to the initial attention-set
even though all the reviewers can proactively add themselves or implicitly
being added when they start interacting with the change.

Bug: Issue 14452
Change-Id: Ie2a6fbc289c66d99986bbaacb6ea4c9864515513