Allow custom initial attention-set from owners
Avoid adding too many reviewers automatically to the attention-set
when a change is created. The OWNERS list, because of hierarchy, may
pull a lot of names as reviewers, but that doesn't mean that all of them
should have the initial task to provide the first feedback.

Expose an API (OwnersAttentionSet that allows an external script
or plugin to specify the logic for selecting the accounts for the
attention set based on:
- Owners
- Change

The new API is called owners-api and needs to be installed
and configured as a libModule.

Bug: Issue 14452
Change-Id: Ie2a6fbc289c66d99986bbaacb6ea4c9864515513