Allow custom initial attention-set from owners
Avoid adding too many reviewers automatically to the attention-set
when a change is created. The OWNERS list, because of hierarchy, may
pull a lot of names as reviewers, but that doesn't mean that all of them
should have the initial task to provide the first feedback.

Expose an API (OwnersAttentionSet that allows an external script
or plugin to specify the logic for selecting the accounts for the
attention set based on:
- Owners
- Change

The new API is contained in owners-api.jar that needs to be installed
and configured as a libModule as follows:

[gerrit]
  installModule = com.googlesource.gerrit.owners.api.OwnersApiModule

Implementing the API for adding owners to attention-set is quite easy
and can be done also through scripting.

In Groovy it would be as easy as:

class MySelector implements OwnersAttentionSet {
  Collection<Account.Id> addToAttentionSet
      (ChangeInfo changeInfo,
       Collection<Account.Id> owners) {
    owners.take(2)
  }
}

For existing owners-autoassign plugins, the injection of the
DynamicItem<OwnersAttentionSet> is optional and therefore is fully
transparent and all owners are added to the attention-set as per
Gerrit default mechanism.

Bug: Issue 14452
Change-Id: Ie2a6fbc289c66d99986bbaacb6ea4c9864515513