Make owners-api.jar lib a mandatory component
Refactor the dependency between owners-autoassign
and owners-api by making the installation of the
owners-api.jar into $GERRIT_SITE/lib explicit and
mandatory.

This simplifies the code and makes the dependency
between the two components cleaner.

From a user's perspective, this is a breaking change
because existing setups would need to add the owners-api.jar
into the $GERRIT_SITE/lib directory, otherwise the
owners-autoassign plugin would fail to load.

Change-Id: I71bd5386a40513fea1f26751b6ccbc95f0d17907