Allow adding reviewers in the OWNERS file
Enable the use of the OWNERS file for adding
extra accounts for being invited automatically to review
without being owners.

Example:

inherited: true
owners:
- some.email@example.com
- User Name
- group/Group of Users
reviewers:
- John Doe
- group/Group of Reviewers
- john@smith.com

NOTE: There is an overlapping of functionality
with the reviewers plugin. The purpose of this extra
support is to avoid duplication of configurations and
painful alignments between the OWNERS and reviewers.config
files.

Bug: Issue 14626
Change-Id: I512e5e73a97fd93b807b25e8ec0597109da6c5cf