Trigger owners-autoassign processing for change readyForReview
The owners-autoassign plugin has historically ignored
all refs/changes/*/meta ref-updates, because they are related
to modifications on the change meta-data.

There is one moment, though, where the evaluation make a lot
of sense, and it is a meta-data only change, triggering only
the ref-update of the refs/changes/*/meta: a WIP change is
set ready for review.

Very often, the user that is working on a WIP change does not
like to have lots of reviewers and therefore tends to remove
them from the reviewers' list. Also, WIP changes do not generate
any attention-set action because the change isn't supposed to
be ready for review.

When the change is later set ready for review *without any
further updates*, the owners-autoassign must process the change
and, if there aren't already the expected owners added,
should add all the owners retrieved from the processing
of the OWNERS files as reviewers.
The detection of the diff cannot rely though on the new SHA1
of the event but rather get the latest patch-set SHA1 and
compute the diff against its change-base.

Bug: Issue 14679
Change-Id: I6b2b6f9b947615f96e331709aa257d54b9c52f07