Only add owners with permissions to attention-set
When auto-assigning reviewers to the change, only the ones that have
visibility on the change itself are added, whilst candidate reviewers
that have no permissions are discarded.

The same logic however was not applied to the attention-set, so that all
candidate owners were blindly added without first checking whether they
actually have permission to read the change.

When such situation happens, an exception is thrown that actually
prevents *all* owners from being added to the attention set.

Only set owners that have read permissions to the attention-set,
similarly to what is already done for reviewers.

Bug: Issue 14728
Change-Id: I46f66369b964b48af953ddbc4597ff19d56d6b23