Auto-assign to CC instead of adding accounts as reviewers
The assignment of all members of the OWNERS file as reviewer
can be seen confusing for certain projects where there is a
large number of devs involved.

Allow to add the accounts as CC instead of being added as
reviewers and leave to the change owner the onus of selecting
which ones of the OWNERS have to be selected as reviewers
and added to the attention-set.

Bug: Issue 15104
Change-Id: I859fbedcec32c6fe0852fd70e30b9a5906e5019f