Adapt to Gerrit v3.{5,6} with the new diff-cache
Gerrit has removed the PatchList cache from v3.6 onwards
and replaced with the new DiffCache.
Avoid using PatchList and use instead the new DiffSummary
coming out of the box from the new DiffCache.

Also add jackson-core explicitly as it isn't provided by
Gerrit anymore and include it as part of the owners plugins.

Change-Id: I99494051eb1dbbb1ca58bbc8e85bdf8a94a3733a