Introduce expandGroups option for disabling groups expansion
Before this change, the owners plugin had always expanded
the groups into the list of accounts included in them.
The expansion generated very long list of names of owners
alongside the associated files in the labels generated.

Introduce a new settings in the owners.config for keeping
the groups unexpanded:

[plugin "owners"]
  expandGroups = false

Keep the groups expansion as default settings for assuring
backward compatibility with the pre-existing behaviour.

Change-Id: Ibf95c3e4377a6370eef069031f4f7330dfb50a8e
TODO: Documentation update and more testing