Avoid using Repository.resolve() when accessing exact ref-names
The Repository.resolve() method is quite complex and expensive
because of the logic of navigating the repository history refs
using arbitrary expressions.

The reading of blobs from the repository for the OWNERS file
does not need any complex navigation logic and is always using
exact refs: use the RefDatabase.exactRef() which uses a direct
lookup which works in this case and is a lot faster.

Change-Id: I6a7ec1ad99f8f549dcbb06599b984c311d6760c0