Introduce GenericMatcher for file-based regex fallback logic
Allow having fallback logic when using the RegexMatchers,
introducing a new specialised subclass called GenericMatcher
which would only be used when none of the other matchers
are selected.

This allows to define rules as the following one:

matchers:
- generic: .*
  owners:
  - group/developers
- regex: .*java
  owners:
  - group/java-developers

A change having the foo.java would then be assigned only
to group/java-developers, because the generic rule would
in theory match but is only used as fallback.

If a change contains two files foo.java and bar.c, the
foo.java will continue to be assigned to the group/java-developers
while the bar.c would be assigned to group/developers as
fallback rule.

Bug: Issue 16480
Change-Id: I2cffb42f72313a660e5f496c84b370a875f34907