Introduce the owners submit rule
Add the SubmitRule implementation that is meant to evaluate change
against the configuration from the OWNERS file. The evaluation is pure
java based.

Note that it kicks in only when `owners.enableSubmitRule = true` and
exits (with log.info) in the other case. At this point it simply prints
the TODO statement.

Bug: Issue 15556
Change-Id: Iea1599f651eeea1dad851ff03fe5bb942106fe39