Check if change is approved by owners
Add simple verification if a change was approved by path owner.
Considering the following OWNERS file configuration:

  inherited: true
  matchers:
  - suffix: extension
    owners:
    - email

Ensure that change that modifies a file with 'extension' is submittable
only when `CR+2` is given by user identified with 'email'.
Note that change that modifies files with different extension is
verified for being submittable too.

Bug: Issue 15556
Change-Id: I8500a484d20c2a79f961f94d761b0322d60a4e65