Introduce the owners submit rule
Add the SubmitRule implementation that is meant to evaluate change
against the configuration from the OWNERS file. The evaluation is pure
java based.

Hote that is hidden behind the `owners.enableSubmitRequirement`
feature flag that needs to be set to `true` in the `gerrit.config`.
At this point only TODO statement is printed to the log.

Bug: Issue 15556
Change-Id: Iea1599f651eeea1dad851ff03fe5bb942106fe39