Ensure that 'Verified' label is handled by default
Add the integration tests that check handling of the 'Verified' label.
Assumptions:
* having 'CR+2' casted by the file owner is not sufficient if 'V+1' is
  missing
* having 'V+1' casted by the file owner is not sufficient if 'CR+2' is
  missing

Note:
In the setup where more than one lable is votable (e.g. where 'Code
Review' and 'Verified' are defined) one needs to configure the owners
plugin which label is binding. This change introduce a method that will
be responsible for getting it from the configuration but for the time
being it is stubbed to always return 'Code Review' (or nothing in case
the 'Code Review' is not present in the setup).

Bug: Issue 15556
Change-Id: I85b989e74f7837365f3c26266bbcfac38e1d5381