[RFC] Add label definition to the OWNERS file
Add possibility to specify which label should be cast by the file
owner. At the moment only one label can be configured for all files but
if `inherited: true` it will be derived from parents.

The following order of definitions is followed (from the highest to the
lowest):
* classic/OWNERS
* OWNERS
* OWNERS in refs/meta/config
* parent's owners definition (following project rules but on parent level)
* n-grand parent's owners definition (following project rules but on
  n-grand parent level) where n is {0 to all the way back to All-Projects}

Example of the OWNERS file with label configuration:

  inherited: true
  label: Code-Review
  owners:
   - user@example.com

If label is configured in OWNERS file then change will be blocked as
long as file owner doesn't vote for it. In current implementation
configuring not existing label results in change being not submittable.

If label is not configured then:
1. if `Code-Review` is configured as required then `Code-Review` is
   expected to be cast by file owner
2. if `Code-Review` is not configured as required then first label that
   is configured as required is expected to be cast by file owner
3. if no label is configured as required but `Code-Review` is
   configured then `Code-Review` is expected to be cast by file owner
4. if no label is configured as required and there is no `Code-Review`
   configured then first label that is configured for project is
   expected to be cast by the file owner

TODO: add documentation once we agree on the implementation

Bug: Issue 15556
Change-Id: Iaaef857d31c5efadfa7016df0ef3208a4c37d87a