Add label definition to the OWNERS file
Add possibility to specify which label should be cast by the file
owner. At the moment only one label can be configured for all files but
if `inherited: true` it will be derived from parents.

The following order of definitions is followed (from the highest to the
lowest):
* classic/OWNERS
* OWNERS
* OWNERS in refs/meta/config
* parent's owners definition (following project rules but on parent level)
* n-grand parent's owners definition (following project rules but on
  n-grand parent level) where n is {0 to all the way back to All-Projects}

Example of the OWNERS file with label configuration:

  inherited: true
  label: Code-Review
  owners:
   - user@example.com

If label is configured in OWNERS file then change will be blocked as
long as file owner doesn't vote for it. In current implementation
configuring not existing label results in change being not submittable.

If label is not configured then:
1. if `Code-Review` is configured then `Code-Review` is expected to be
   cast by file owner
2. if there is no `Code-Review` label configured it will be assumed that
   wrong label was configured and submission will be blocked

Note that the following scenarios are not yet implemented in owners
submit requirement:
* configuring group of owners
* making `+1` from owners sufficient for label that requires `+2` by default
  (e.g. Code-Review)
* displaying owners details on a pre-file basis, the `submit_records`
  entry contains the following error message

    "error_message":"Missing approvals for path(s): [README.md]"

  but it is not displayed in the Gerrit's UI (as of `stable-3.5`)

Bug: Issue 15556
Change-Id: Iaaef857d31c5efadfa7016df0ef3208a4c37d87a