Add label score definition to the OWNERS file
That allows to configure a specific score value that is sufficient from
the OWNERS perspective to approve the file changes.

For instance the following OWNERS configuration:

  inherited: true
  label: Code-Review, 1
  owners:
  - group/Legal
  matchers:
  - exact: LICENSE

Means that in case when LICENSE file gets modified then someone from the
legal department has to give `Code-Review+1`. IOW `Code-Review+2`
doesn't have to granted to all project files.

Notes:
* label score definition is global for this file and applies to all
  matching files (or files if there are not matchers specified)
* when the `inherited: true` and no label definition then it is
  inherited from the parent projects (if specified)
* regular developers still have to give `Code-Review+2` for a
  change to be submittable as default submit rule is still evaluated

Bug: Issue 15556
Change-Id: I76627192b8627139a04e1feb6eb13109d8f3717d