Cache the OWNERS file's parsed content
The OWNERS file is rarerly modified however it is being read and parsed
for every single change that gets evaluated. Introduce the cache (called
`owners-path_owners_entries`) that holds the PathOwnersEntry value for a
project and branch (typically `refs/meta/config` and development
branch). It is invalidated when project's ref is updated.
As a result it reduces 2-3 times the configuration load latency.

There are several constraints that influenced the implementation
* the `PathOwners` is a common class used by both owners and
  owners-autoassign plugins
* the `PathOwners`, in the owners plugin prolog's engine, is
  instantiated by the constructor call and injection is not possible
  there

Bug: Issue 16830
Change-Id: Ic7d61de07f3ec5169f0cf7c4ae8266351543e108