Follow the whole chain of parent projects for OWNERS
The OWNERS file can be stored in refs/meta/config for being
used as global fallback. However, when following the project
hierarchy it was only considering the 1st-level parent and not
all the other parents up the chain.

Get the full chain of parents and use it for the processing
of the equivalent PathOwners to be used in the Prolog rules
evaluation.

Bug: Issue 16893
Change-Id: I9bf46bcdf977e5645d6b44c4f08a792fdf3a4230