Make code_review_user/1 a native plugin predicate
The code_review_user/1 predicate needs to have access to
the LabelTypes defined in the project through Java API for
making sure to consider the real maximum value of a label
instead of having +2 hardcoded.

Previously the predicate was hardcoded to checking for
a Code-Review +2 which may not work if the maximum label
score is different than +2.

Bug: Issue 16970
Change-Id: Iad61cb96a9c894a69537e79e91cb094d959b0cc8