Fix `OWNERS` file over-caching problem
Issue:
When 'OWENRS' file of the parent's refs/meta/config is modified after
child's own file the updates are not reflected in the child's PathOwners
content.

Solution:
Cache only the parsed 'OWNERS' file content so that it doesn't have to
be read over and over again from the disk however parent's hierarchy is
still traversed. In addition make sure that static EMPTY entry is only
used for reading.

Bug: Issue 16830
Change-Id: I6eed6cee12ae99ed95387fb1518129dc3aa1101b