Filter out the files owners based on the owners' code-review value
The GetFilesOwners did not process the ownership's review labels for
filtering out the total set of files owners requirements, making the
list of owners' requirements less effective.

The GetFilesOwners should not display anymore the files that have their
owners requirements already satisfied. Once all the owners have provided
their feedback, the payload of GetFilesOwners, limited to the files
list, it should be empty.

Bug: Issue 40015590
Change-Id: I7b513985c0711d9e3c318e629f8529effcd1a190