Clean up `OwnersSubmitRequirement` class
The `OwnersSubmitRequirement` doesn't have implement `SubmitRule`
anymore therefore it can be simplified. The following changes were
performed:
* the `evaluate` method returns `boolean` that is expected by a calling
  class
* the `evaluate` method throws `SubmitRequirementEvaluationException`
  when submit requirements cannot be evaluated
* `SubmitRule` related functions and constants were removed

As a result `evaluate` method was simplified and simply checks if any
owner's vote is missing. The calling class was also simplified as
returned value doesn't have to be adjusted to
`SubmitRequirementPredicate` interface.

Notes:
The `OwnersApprovalHasPredicate` and `OwnersSubmitRequirement` could be
merge into one class but:
* proposed solution is less intrusive therefore it is easier to review
  (note that the corresponding unit tests were not modified)
* it will be easier to apply any fixes in case when they were applied to
  `stable-3.5` - they have high chance to be just merged up otherwise
  they will have to re-applied

Bug: Issue 345837787
Change-Id: Idaed3604aeae4e6c3f2ca1ef07f2905e4ac30a01