Always expose `has:approval_owners` submit requirement predicate
The `owners.enableSubmitRequirement = true` is a global switch that
allows for quick site switch from `prolog` submit rule to plugin's
provided submit rule. However in case when such migration is not
feasible (e.g. when `prolog` rules need to be slowly phased out or when
more control is needed over rule's applicability) one can now turn
submit requirement in `project.config` for a certain project (or
hierarchy of projects), without turning it on globally.

Changes:
* the `approval_owners` submit requirement operand is always registered
* all existing integration tests were adjusted to verify that plugin
  works correctly also n new mode
* the configuration was updated to highlight such possibility

Bug: Issue 345837787
Change-Id: Iaca45dd47e23d5f2d40a2732b52abdaeba170e8d