Use simple clone instead of registerRepoConnection()
Gerrit master does not allow to access the registerRepoConnection()
for allowing tests to access the in-memory repository push URL.

Use a simple clone() of the repository which would automatically
manage the in-memory repository use-case. Also fetch `refs/meta/config`
and reset repo to it so that tests can use it.

Change-Id: I6582ed4335985938d06f2117a82a97a58c0cec99