Clarify that examples are using a default labels configuration
By reading the owners plugin documentation it looks like some
defaults are invarant, like the Code-Review label and its score.
However, when customising Gerrit default Code-Review label, the
behaviour may change radically depending on the changes made
and the owners plugin documentation would therefore become misleading.

It is an impossible task to document all the possible combinations
of the label settings and the implied behaviours on the owners plugin,
therefore just clarify the context of the examples so that the
reader would be advised on the underlying preconditions.

Change-Id: Iebe702b7956ba275140f882ac3bc0750cb5918fd