Restore the ability for global definition of the owners' label
When moving from the Prolog-style to the Submit Requirements-style
definition of the owners, the ability to enforce a specific label
for all projects, and OWNERS file was lost due to the fact that
the child projects and OWNERS files inside could have just avoided
to inherit from their parents, skipping the enforcement of a global
label.

With Prolog it was instead possible to define a global rules.pl
that would have enforced a specific label, regardless of what
the individual projects would have been defined in their OWNERS files.

Introduce a new 'owners.label' settings in the plugin config
owners.config so that the Gerrit admin can enforce a global
label regardless of the will of every project and its hierarchy.

By default, the global label definition is undefined and, therefore
this change is fully backward compatible with the existing
setups.

Change-Id: I19ae0e9029338705214fd447defe6ba6f7ec6966