Introduce `allFilesApproved` and `filesOwners` states to plugin UI
The `OwnersModel` was extended with this two states. They are loaded
reactively from plugin through the `OwnersService`. Both header and
content UI elements subscribe on these states updates and basing on that
update their corresponding internal states. As a result `render` method
is called for each element. At the moment values of `allFilesApproved`
and `filesOwners` are logged to the console.

Notes:
* UI elements have `change` property that gets updated by Gerrit UI when
  update is detected while details screen is opened, as a result new
  model and model loader are created, UI elements subscriptions are
  refreshed and new state values are loaded, leading eventually to
  reacitve UI refresh
* despite `allFilesApproved` logic that prevents from unecessary calls
  to obtain file owners it is still called for each file (as they form
  independent UI elements). It will be further improved (in a later
  follow up change) with an introducetion of cache (between service and
  plugin api).

Change-Id: I1ac079a293033797c3b8dd1c06581d8a0c10fba7