Disaply file owner's vote value in the File Owners summary
Display the vote value of a label that is required in order to satisfy
the OWNERS definition as returned from the plugin's API (typically
`Code-Review`). Note that default or different (e.g. `Verified`) labels
are not shown as they are not relevant from the owners perspective.

TODO: add copy file owner's email button

Bug: Issue 373151160
Change-Id: I07919c66f4cda95da40a62086a1b832aec7e8a4a