Display `Copy owner's email` button for owner
When owners list is returned from the plugin's REST API it contains
only the account id. Email is available in account details therefore
it can be obtained from the Gerrit's REST API. But that:
* would mean that accout details are obtained two times - once for the
  change screen purposes and another time for plugin
* `gr-account-label` or similar would have to be implemented from
  scratch

In order to circumvent that the alternative approach was chosen. The
`gr-account-label` component (which is a part of `gr-owner` element)
contains all the necessary means to obtain account details therefore:
* when the initial `gr-owner` element updsate is performed
  `gr-account-label` instance is obtained (as defined by `GrAccountLabel`
  interface)
* that instance is used to fill account details - note that it leverages
  the caching abilities of the `AccountState` therefore given account
  details are obtained only once

Bug: Issue 373151160
Change-Id: I8dcffa067e07ecc7e63df6d12bca35d1791bd730