Introduce `Owned Files` tab to the change screen
This is the initial change that introduced the `Owned Files` tab to the
change screen. At the moment tab is shown only when user owns any file
(both `ownedFiles` and `shouldHide` functions are covered in units) and
change is not in the `EDIT` mode.

All files are displayed in the tab content (TODO provide links to diffs
and style it appropriately).

Changes:
* `change-view-tab=[header|content]` were used to register new plugin
  components
* `getLoggedInUserRole` service's function was modified to
  `getLoggedInUser` so that user account and role are returned (existing
  functionality and tests were adjusted)
* `getRandom` function was moved to `test-utils.ts` so that it gets
  re-used (note that function is available only in tests)

Bug: Issue 376837413
Change-Id: Ica25f84596d0d7d2dfe268e97a4e3477ce605a5f