Introduce OwnersApiCache
When Gerrit UI loads files owners status it treats each file as
an independent UI component therefore as a result plugin's REST API
endpoint is called for each one of them. The `Owned Files` tab adds an
extra call on the top of that. The same applies for getting the
currently logged in user account details.

Reduce it to a single call per REST API endpoint by OwnersApiCache
introduction (the idea is again transplanted from `code-owners` plugin).

It is a safe bet as in the event of change modification (for instance
when one votes) Gerrit doesn't modify the existing change object. It
creates new change object instead and existing
`OwnersService.getOwnersService` function (that is used to obtain the
service) ensures that new service instance gets created (together with
new instance of `OwnersApiCache`).

Caching abilities are also verified with unit tests.

Bug: Issue 377723880
Change-Id: I4b7312b644a0d2a608e4f19e23eda4d5a87f0b51